Made DeserializeResponseBody asynchronous
Issue #455 asked to add `ConfigureAwait(false)` to the  DeserializeResponseBody's call to ReadAsStringAsync. Since that call was synchronous (with Result) the only reason to add the ConfigureAwait(false) was to make it async. This means also changing the method's name to be coherent with the rest of the library.

Fixes issue #455